refactor: moved all schematics code to own file
